backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
drf-yasg>=1.21.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
Pillow>=10.0.0
gunicorn>=21.0.0
whitenoise>=6.5.0